move the network check to the health check task:
- remove the standalone perf task network-health and add it as a check
  in the healthcheck task.
- modify the schedule of the healthcheck to be every 15 instead of 20
  mins
- modify the report of the network check to only report the failed
  services for a better alignment with the other healthcheck
