+++ cmds/modules/noded/main.go
@@ -25,7 +25,6 @@ import (
 	"github.com/threefoldtech/zos/pkg/perf/cpubench"
 	"github.com/threefoldtech/zos/pkg/perf/healthcheck"
 	"github.com/threefoldtech/zos/pkg/perf/iperf"
-	"github.com/threefoldtech/zos/pkg/perf/networkhealth"
 	"github.com/threefoldtech/zos/pkg/perf/publicip"
 	"github.com/threefoldtech/zos/pkg/registrar"
 	"github.com/threefoldtech/zos/pkg/stubs"
@@ -220,7 +219,6 @@ func action(cli *cli.Context) error {
 	perfMon.AddTask(cpubench.NewTask())
 	perfMon.AddTask(publicip.NewTask())
 	perfMon.AddTask(healthcheck.NewTask())
-	perfMon.AddTask(networkhealth.NewTask())
 
 	if err = perfMon.Run(ctx); err != nil {
 		return errors.Wrap(err, "failed to run the scheduler")
